+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Type of instance.Version was changed from string to System.Version.
+
 ## [0.2.0] - 2022-03-27
 
 ### Added